# Library/Swagger/docs/swagger.yaml
basePath: /employee
definitions:
...
  Response.RespErrorStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        example: 400
        type: integer
      title:
        example: bad_request
        type: string
    type: object
  Response.RespResultStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: 'Example: Bearer abcdefghijklmnopqrstuvwxyz12345678901'
  title: Employee Service
  version: "1.0"
paths:
...
  /dashboards/employee/length-of-work/{numberOfYear}:
    get:
      consumes:
      - application/json
      parameters:
      - description: numberOfYear
        in: path
        name: numberOfYear
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TotalEmployeeByLengthOfWork
      tags:
      - Dashboards
  /dashboards/employee/new:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: NewEmployeeData
      tags:
      - Dashboards
  /dashboards/employee/religion/summary:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TotalReligionSummary
      tags:
      - Dashboards
  /dashboards/employee/work-status/summary:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TotalWorkStatusSummary
      tags:
      - Dashboards
  /dashboards/employee/work-status/will-expire/summary:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: TotalWillExpireEmployeeContract
      tags:
      - Dashboards
...
securityDefinitions:
  Bearer:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"